backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
pandas==2.2.3
pyarrow==19.0.1
python-dotenv==1.0.1